backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
python-dotenv
Pillow
